vilatec/requirements.txt
aiohttp==3.7.2
appdirs==1.4.3
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==20.3.0
automationhat==0.2.0
beautifulsoup4==4.7.1
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
cachetools==4.1.1
Cap1xxx==0.1.3
certifi==2018.8.24
chardet==3.0.4
Click==7.0
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==2.6.1
cycler==0.10.0
decorator==4.3.0
dnspython==1.16.0
dominate==2.6.0
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
eventlet==0.29.1
ExplorerHAT==0.4.2
Flask==1.0.2
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
fourletterphat==0.1.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gpiozero==1.6.2
greenlet==0.4.17
guizero==0.6.0
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.18.1
idna==2.6
ipykernel==4.9.0
ipython==5.8.0
ipython-genutils==0.2.0
iso8601==0.1.13
itsdangerous==0.24
Jinja2==2.10
jupyter-client==5.2.3
jupyter-core==4.4.0
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.0.1
lxml==4.3.2
MarkupSafe==1.1.0
matplotlib==3.0.2
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
multidict==5.0.0
names==0.3.0
nudatus==0.0.4
numpy==1.16.2
oauth2client==4.1.3
oauthlib==2.1.0
observable==1.0.3
olefile==0.46
paho-mqtt==1.5.1
pantilthat==0.0.7
pexpect==4.6.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
pickleshare==0.7.5
picraft==1.0
piglow==1.2.5
pigpio==1.78
Pillow==5.4.1
prompt-toolkit==1.0.15
protobuf==3.13.0
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.4.0
pycrypto==2.6.1
pyflakes==2.0.0
pygame==1.9.4.post1
Pygments==2.3.1
PyGObject==3.30.4
pygsheets==2.0.3.1
pyinotify==0.9.6
PyJWT==1.7.0
pyOpenSSL==19.0.0
pyparsing==2.2.0
pyserial==3.4
python-apt==1.8.4.3
python-dateutil==2.7.3
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.4
pyxdg==0.25
pyzmq==17.1.2
qtconsole==4.3.1
rainbowhat==0.1.0
repackage==0.7.3
requests==2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
RPi.GPIO==0.7.0
rsa==4.6
RTIMULib==7.2.1
schedule==0.6.0
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
semver==2.0.1
sense-hat==2.4.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.15.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==1.8
spidev==3.5
SQLAlchemy==1.3.20
ssh-import-id==5.7
telepot==12.7
tornado==5.1.1
touchphat==0.0.1
traitlets==4.3.2
twython==3.7.0
typing-extensions==3.7.4.3
uflash==1.2.4
unicornhathd==0.0.4
uritemplate==3.0.1
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
yarl==1.6.2
